[Fix] Change app to use Blueprints. Rearrange folder structure.
