+++ .gitignore
@@ -3,6 +3,7 @@ __pycache__/
 .pytest_cache/
 *.py[cod]
 *$py.class
+.coveragerc
 
 # C extensions
 *.so
@@ -40,7 +41,7 @@ pip-delete-this-directory.txt
 htmlcov/
 .tox/
 .coverage
-.coverage.*
+.coverage*
 .cache
 nosetests.xml
 coverage.xml
