+++ .travis.yml
@@ -7,6 +7,6 @@ install:
   - pip3 install coveralls
 # command to run tests
 script:
-  - nosetests --with-coverage  # or py.test for Python versions 3.5 and below
+  - nosetests --with-coverage --cover-package=app  # or py.test for Python versions 3.5 and below
 after_success:
   - coveralls
